+++ src/Altinn.Correspondence.Application/MigrateCorrespondence/MigrateCorrespondenceResponse.cs
@@ -4,7 +4,7 @@ public class MigrateCorrespondenceResponse
 {
     public Guid CorrespondenceId { get; set; }
 
-    public long Altinn2CorrespondenceId { get; set; }
+    public int Altinn2CorrespondenceId { get; set; }
 
     public List<AttachmentMigrationStatus>? AttachmentMigrationStatuses { get; set; }
 }
